Introduce a Xcode project to build correctly
