+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "test": "xo",
-    "build": "swiftc swift/Recorder.swift swift/main.swift -o swift/main",
+    "build": "cd swift && xcodebuild && mv build/release/aperture main && rm -r build",
     "postinstall": "npm run build"
   },
   "repository": {
